move PoolSizes from src/internal to src/scheduling
