+++ kotlinx-coroutines-core/jvm/src/scheduling/PoolSizes.kt
@@ -1,0 +1,4 @@
+package kotlinx.coroutines.scheduling
+
+internal actual val CONCURRENT_CORE_POOL_SIZE: Int = CORE_POOL_SIZE
+internal actual val CONCURRENT_MAX_POOL_SIZE: Int = MAX_POOL_SIZE
