+++ kotlinx-coroutines-core/native/src/scheduling/PoolSizes.kt
@@ -1,4 +1,4 @@
-package kotlinx.coroutines.internal
+package kotlinx.coroutines.scheduling
 
 internal actual val CONCURRENT_CORE_POOL_SIZE: Int = Platform.getAvailableProcessors().coerceAtLeast(2)
 internal actual val CONCURRENT_MAX_POOL_SIZE: Int = 2048
